- fixed 64 bit compile warning - less aggressive logging for phones that don't support option methods in dlg
git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/trunk@7957 689a6050-402a-0410-94f2-e92a70836424

+++ modules/dialog/dialog.c
@@ -424,7 +424,7 @@ static int fixup_create_dlg2(void **param, int param_no)
 		}
 	}
 
-	*param=(void *)(unsigned int)flags;
+	*param=(void *)(unsigned long)flags;
 
 	return 0;
 }
